requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==9.1.0
packaging==23.2
pyproject_hooks==1.0.0
pytz==2023.3.post1
requests==2.31.0
urllib3==2.1.0
waitress==3.0.0
Werkzeug==3.0.1
